+++ package.json
@@ -53,6 +53,8 @@
     "copy-to-clipboard": "^3.3.1",
     "react": "^17.0.2",
     "react-codemirror2": "^7.2.1",
-    "react-dom": "^17.0.2"
+    "react-dom": "^17.0.2",
+    "react-download-link": "^2.3.0",
+    "react-downloader-file": "^1.0.0"
   }
 }
