+++ src/app/components/Editor.css
@@ -1,0 +1,5 @@
+.CodeMirror {
+  height: auto;
+  border-radius: 5px;
+  font-size: 12px;
+}
