--- src/backend/testfiles/test.txt
@@ -1,8 +1,0 @@
-// Initial describe statement for default initialized state
-  describe('default state', () => {
-    it('should return a default state when given an undefined input', () => {
-      // expect(currMemoizedState[0]).toEqual({});
-      // expect(currMemoizedState[1]).toEqual({});
-      // expect(currMemoizedState).toEqual([{},{}]);
-    });
-  });
