+++ src/extension/content.js
@@ -18,15 +18,21 @@ function handleMessage(request, sender, sendResponse) {
     console.log('in content.js, request.data: ', request.data)
     // Send a message to background.js with the shape { action: 'addTest', message: [(testArray)] }
     return chrome.runtime.sendMessage({ action: request.data.type, message: request.data.message });
-  }
+  } 
   // if (request.data.type === 'initial panel load'){ 
   //   console.log('inside initial panel load in content.js')
   //   return chrome.runtime.sendMessage({ action: request.data.type});
   // }
 }
 
-// Listening for me sseges in the window, if it recieves a 'message', invoke handleMessage()
+// Window Listening for messeges in the window, if it recieves a 'message'
 window.addEventListener('message', handleMessage);
+
+// Chrome Listening for messeges in the browser, if it recieves a 'message'
+chrome.runtime.onMessage.addListener((request) => {
+  // Send a postMessage to window to forward request to injected.js
+  window.postMessage(request, '*');
+})
 
 // Send a message to background.js with the shape { action: 'injectScript' } to injectScript
 chrome.runtime.sendMessage({ action: 'injectScript' })
